Apply power-profiles-daemon.service preset on upgrades to F35 and F36
The preset was changed to enabled in F35, but systemd rpm scriptlets
don't have a mechanism for re-applying presets when the defaults were
changed.

This commit does it manually instead and repurposes an older trigger
scriptlet to re-apply the preset on distro upgrades.

I've changed the existing trigger a bit to match what we came up with
for enabling fstrim.timer in util-linux a while back. This should be
a well tested and a known working trigger.